requirements.txt
flask
flask-socketio
python-socketio
eventlet
opencv-python-headless
ultralytics
numpy
geopy